chapter3/pyproject.toml
[tool.poetry]
name = "aws-container"
version = "0.1.0"
description = ""
authors = ["m4-sato <mssst1116@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11" # 特定のPythonバージョンを指定
boto3 = "1.34.87"
langchain-aws = "0.1.4"
langchain-community = "0.2.0"
python-dateutil = "2.8.2"
python-dotenv = "^1.0.0"
langchain = "^0.2.0"  # 最新バージョンを使用
langsmith = "^0.1.17"  # langchainが要求する範囲内のバージョンを使用
langchain-core = "^0.2.0"
streamlit = "^1.33.0" 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
